+++ src/editor/core/plugins/resize_nav/menu.js
@@ -5,9 +5,8 @@
  *
  */
 
-import {icons, MenuItem} from "../../menu/menu"
-
-const SELECTOR_DEFAULT = '.helper-group, .format-group, .insert-dropdown, .ProseMirror-menu-insertTable:not(.hidden), .ProseMirror-menu-fullScreen:not(.hidden)';
+import {icons, MenuItem} from "../../menu"
+import {toggleNavPluginKey, isCollapsed} from "./plugin"
 
 function resizeNav(context) {
 
@@ -15,16 +14,14 @@ function resizeNav(context) {
         id: 'resizeNav',
         title: "More",
         sortOrder: 400,
-        run() {
-            let $nodes = getNodes(context);
-            if(!context.editor.$.find('.helper-group').is(':visible')) {
-                $nodes.fadeIn();
+        run(state, dispatch) {
+            let wasCollapsed = isCollapsed(state);
+            dispatch(state.tr.setMeta(toggleNavPluginKey, !wasCollapsed));
+
+            if(wasCollapsed) {
                 this.switchIcon(icons.angleDoubleLeft);
-                $(this.dom).data('state', true);
             } else {
-                $nodes.hide();
                 this.switchIcon(icons.angleDoubleRight);
-                $(this.dom).data('state', false);
             }
         },
         icon: icons.angleDoubleRight
@@ -47,4 +44,18 @@ export function menu(context) {
             item: resizeNav(context)
         },
     ]
+}
+
+export function menuWrapper(context) {
+    return {
+        select: function(menuItem, state, active) {
+            let collapsed = toggleNavPluginKey.getState(state);
+
+            if(collapsed && menuItem.options.hideOnCollapse) {
+                return false;
+            }
+
+            return active;
+        }
+    }
 }