+++ src/editor/core/plugins/resize_nav/plugin.js
@@ -1,0 +1,22 @@
+import {Plugin, PluginKey} from "prosemirror-state";
+
+export const toggleNavPluginKey = new PluginKey('toggleNav');
+
+export function isCollapsed(state) {
+    return toggleNavPluginKey.getState(state);
+}
+
+export function toggleNavPlugin(context) {
+    return new Plugin({
+        key: toggleNavPluginKey,
+        state: {
+            init(config, state) {
+                return true;
+            },
+            apply(tr, prevPluginState, oldState, newState) {
+                let meta = tr.getMeta(toggleNavPluginKey);
+                return typeof meta !== 'undefined' ? meta : prevPluginState;
+            }
+        },
+    })
+}