+++ src/editor/core/plugins/source/index.js
@@ -4,50 +4,13 @@
  * @license https://www.humhub.com/licences
  *
  */
-import {menu} from "./menu"
-import {sourcePlugin, isSourceMode} from "./plugin"
-
-const enabledItems = [
-    'source',
-    'resizeNav',
-    'fullScreen'
-]
+import {menu, menuWrapper} from "./menu"
+import {sourcePlugin} from "./plugin"
 
 const source = {
     id: 'source',
     menu: (context) => menu(context),
-    menuWrapper: (context) => {
-      return {
-          run: function(menuItem, state) {
-              if(menuItem.options.id === 'source' || !isSourceMode(state)) {
-                  return false;
-              }
-
-              if(menuItem.runSource) {
-                  menuItem.runSource();
-                  return true;
-              }
-
-              return false;
-          },
-          enable: function(menuItem, state, enabled) {
-              let sourceMode = isSourceMode(state);
-
-              if(enabledItems.includes(menuItem.options.id) || (sourceMode &&  menuItem.runSource)) {
-                  return true;
-              }
-
-              return sourceMode ? false : enabled;
-          },
-          active: function(menuItem, state, active) {
-              if(menuItem.options.id === 'source') {
-                  return active;
-              }
-
-              return (isSourceMode(state)) ? false : active;
-          }
-      }
-    },
+    menuWrapper: (context) => menuWrapper(context),
     plugins: (context) => {
         return [
             sourcePlugin(context)